requirements.txt
asgiref==3.7.2
cfgv==3.4.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==4.2.7
django-import-export==3.3.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.13.1
identify==2.5.32
isort==5.12.0
MarkupPy==1.14
nodeenv==1.8.0
odfpy==1.4.1
openpyxl==3.1.2
platformdirs==4.0.0
pre-commit==3.5.0
pytz==2023.3.post1
PyYAML==6.0.1
ruff==0.1.6
sqlparse==0.4.4
tablib==3.5.0
virtualenv==20.24.7
xlrd==2.0.1
xlwt==1.3.0
